+++ libs/rxjs/array/src/lib/from-object-keys.ts
@@ -9,7 +9,9 @@ import { map } from 'rxjs/operators';
  * Returns an Observable that emits an array from a source `Object` using Object.keys, the array contains
  * the object keys as strings.
  *
- * @category Object
+ * @category Create
+ *
+ * @remarks Regardless of Object key type the result Array will have a `string` key value
  *
  * @typeParam K The key type of the source Object
  * @typeParam T The value type of the source Object
@@ -21,7 +23,7 @@ import { map } from 'rxjs/operators';
  * ```
  * Output: `[ '1', '2', '3' ]`
  *
- * @returns Observable that emits a Array from a source Object keys
+ * @returns Observable that emits a Array of strings from a source Object keys
  */
 export function fromObjectKeys<K extends string | number | symbol, T extends unknown>(): OperatorFunction<
   Record<K, T>,