+++ libs/rxjs/array/src/lib/last-index-of.ts
@@ -4,17 +4,17 @@
  */
 import { OperatorFunction } from 'rxjs';
 import { map } from 'rxjs/operators';
-import { MapFn } from '../types/generic-methods';
 
 /**
- * Returns an Observable number or array of numbers. These are the index numbers of first truthy value in the source array
- * using Array.lastIndexOf
+ * Returns an Observable number or array of numbers. These are the index numbers of first truthy value in the source
+ * array using Array.lastIndexOf
  *
  * @category Query
  *
+ * @typeParam T The input type of the source Array or Set
+ *
  * @param input A value or array of values to get the index of in the source array
  * @param fromIndex Optional index to start searching from in the array
- * @param mapFn Optional [[MapFn]] that can be used to make comparison easier (such as lower casing text)
  *
  * @example
  * Returns the last index of the word `RxJS` in the array
@@ -51,21 +51,16 @@ import { MapFn } from '../types/generic-methods';
  * @returns Observable number or array of numbers containing the index of the last found value
  */
 export function lastIndexOf<T extends unknown>(
-  input: T | T[],
+  input: T | T[] | Set<T>,
   fromIndex?: number,
-  mapFn?: MapFn<T>,
-): OperatorFunction<T[], number | number[]> {
+): OperatorFunction<T[] | Set<T>, number | number[]> {
   return (source) =>
     source.pipe(
-      map((value) => {
+      map(([...value]) => {
         fromIndex = fromIndex || value.length - 1;
-        return Array.isArray(input)
-          ? (input.map((inputVal) =>
-              mapFn ? value.map(mapFn).lastIndexOf(inputVal, fromIndex) : value.lastIndexOf(inputVal, fromIndex),
-            ) as number[])
-          : mapFn
-          ? value.map(mapFn).lastIndexOf(input, fromIndex)
-          : (value.lastIndexOf(input, fromIndex) as number);
+        return Array.isArray(input) || input instanceof Set
+          ? [...input].map((inputVal) => value.lastIndexOf(inputVal, fromIndex))
+          : value.lastIndexOf(input, fromIndex);
       }),
     );
 }