+++ libs/rxjs/array/src/lib/reverse.ts
@@ -3,13 +3,15 @@
  * @module Array
  */
 
-import { Observable, OperatorFunction } from 'rxjs';
+import { OperatorFunction } from 'rxjs';
 import { map } from 'rxjs/operators';
 
 /**
  * Returns an Observable that emits array taking the source and running the result of Array.reverse
  *
  * @category Modify
+ *
+ * @typeParam T Item type contained in the Array/Set
  *
  * @example
  * Reverse an array of values
@@ -21,6 +23,6 @@ import { map } from 'rxjs/operators';
  *
  * @returns Observable that emits an array which is reversed from the source array
  */
-export function reverse<T extends unknown>(): OperatorFunction<T[], T[]> {
-  return (source: Observable<T[]>) => source.pipe(map((value) => [...value].reverse()));
+export function reverse<T extends unknown>(): OperatorFunction<T[] | Set<T>, T[]> {
+  return (source) => source.pipe(map((value) => [...value].reverse()));
 }